app/src/main/java/com/pr656d/cattlenotes/ui/launch/LaunchViewModel.kt
package com.pr656d.cattlenotes.ui.launch

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import com.pr656d.cattlenotes.data.local.prefs.PreferenceStorage
import com.pr656d.cattlenotes.ui.launch.LaunchViewModel.LaunchDestination.LOGIN_ACTIVITY
import com.pr656d.cattlenotes.ui.launch.LaunchViewModel.LaunchDestination.MAIN_ACTIVITY
import com.pr656d.cattlenotes.utils.Event
import javax.inject.Inject

class LaunchViewModel @Inject constructor(sharedPreferenceStorage: PreferenceStorage) : ViewModel() {

    private val _launchDestination = MutableLiveData<Event<LaunchDestination>>()
    val launchDestination: LiveData<Event<LaunchDestination>> = _launchDestination

    init {
        _launchDestination.value = Event(
            if (sharedPreferenceStorage.loginCompleted)
                MAIN_ACTIVITY
            else
                LOGIN_ACTIVITY
        )
    }

    enum class LaunchDestination {
        MAIN_ACTIVITY,
        LOGIN_ACTIVITY
    }
}